// src/app/employees/interfaces/empdata.interface.ts
// Generated by https://quicktype.io

export interface EmpData {
  d: D;
}

export interface D {
  results: Result[];
}

export interface Result {
  __metadata?:         Metadata;
  Compliance:         string;
  Countryname:        string;
  EmpId:              string;
  Mandt?:              string;
  Name:               string;
  Address:            string;
  Salary:             string;
  Curcode:            string;
  EmpDataDetailNavig?: EmpDataNavig;
  EmpDataSkillsNavig?: EmpDataNavig;
}

export interface EmpDataNavig {
  __deferred: Deferred;
}

export interface Deferred {
  uri: string;
}

export interface Metadata {
  id:   string;
  uri:  string;
  type: string;
}